completed adjustments to fix OOIION-631
