+++ ion/agents/platform/notes.txt
@@ -1,9 +1,13 @@
 platform agent framework
 
+some noteworthy notes (occasionaly updated)
 
-2013-01-09
+2013-01-11
 - OOIION-631 Use system time (get_ion_ts and similar) instead of NTP
   - get_resource and set_resource adjusted.
+  - refactor: moved *DriverEvent classes from platform_driver into new
+    module platform_driver_event.
+  - Removed the unneeded ts attribute in *DriverEvent classes.
 
 2012-10-31:
 - preparations for IOC demo
